Avoid throwing frontend javascript errors (#279)
By only looking for err.responseJSON.redirectUrl, you're assuming that an ajax call will never return a status code of 500. If such an action happens, as undesirable as that may be, we don't also want to trigger a front-end error.

Personally, this change will allow my global event handlers ($(document).ajaxError() and $(document).ajaxComplete(), for example) to fire successfully.