Pass SFRA store model, not demandware Store object (#439)
At the moment, storeDetails.isml has to work with both the models/store.js object and Demandware's Store object. This is possible in base SFRA because all of the field names are identical. However, changes made to the store model by end users will be inaccessible in the in-map version of storeDetails.isml.

The change above should cause no regression issues because of attribute name re-use between Demandware and SFRA.