Increase extensibility of shipping.js (#400)
This is meant to be a starting point for extending front-end modules.

I should start by explaining the why of this.  shipping.js, like most front-end JS in the cartridge, exports an extensible module.  It exports both top level methods to be run with processIncludes, as well as a methods object containing methods that shouldn't be run there but should still be accessible.  As I understand it, one of the main reasons behind this is to allow the module to be added to without modifying base.  It also allows for individual methods to be overwritten as needed without having to modify base or having to overwrite the entire module.  The problem is that the top level methods call functions that are scoped to the file.  So let's say I want to overwrite just the method "selectShippingMethodAjax."  I would do this by defining the new function inside of a file in my cartridge.  I pull in the base shipping module, and then export a version of it with "methods.selectShippingMethodAjax" set to my new function.  Now, I would like this change to affect wherever "selectShippingMethodAjax" is called.  An example is "selectShippingMethod."  But it won't use my updated method.  It will use the one defined in base.  And thus, I also have to overwrite "selectShippingMethod" for my changes to take affect.  Thus, in order to update one function, I have to overwrite two.  This creates more technical debt than is ideal.  

Now for the fix.  What this does is, in each top level function, it tries to get the module at time the function is run as "baseObj."  It then checks that the module has an exported version of the function we want to run.  If it does, it runs that.  Thus, it runs the most up-to-date version of that function.  And if it doesn't, for any reason, it falls back to the version defined in the file.