Updating server.js to support calling a hook for route registration (#648)
* Updating server.js to support calling a hook for route registration

Adding hook `app.server.registerRoute`. Executes when a new route is registered in SFRA. Passes the route, which allows hook implementations to then register route events for global processing.

Note: There doesn't appear to be a way to determine which controller this is executing for, only the route name. Identifying/passing the controller could be something else useful (should be stored on the Route itself imho). In the meantime, identifying current requested page can be used via request.httpPath (or req.path), but this couldn't be depended upon in the scenario of a controller calling yet another controller (meaning the subsequent controller would not match the httpPath that was identified).

* Added mock for HookMgr

* Fixing lint errors