added to render the home page isml template in Default-Start, as that… (#729) (#738)
* added to render the home page isml template in Default-Start, as that controller is called when sitepath is defined in the site aliases
as per the documentation Was the matching hostName and sitePath (if any) found in the settings property?
Yes - examine the remaining portion of the entered URL (the portion after the hostName and sitePath, if any) to determine the correct locale and nature of the requested resource (that is, to determine whether the URL specifies a category, content asset, or pipeline). If the entered URL specifies only the hostName and sitePath, or only the hostName, execute the Default-Start pipeline and return the response to the customer. Otherwise, retrieve the requested resource (executing a pipeline if necessary) and return the requested resource to the customer. If no locale is specified, use the default locale for the site when retrieving the requested resource. Skip all of the remaining steps in this process. - which is causing infinite redirect when combined with "entry-point-pipelines" : [ "Default-Start", "Home-Show" ]

* fixed the controller start node