Fix null reference error in order confirmation mail (#737)
The model built for the order confirmation mail lacks
some properties that the template tries to render.

It seems that the order line items are being decorated with `priceTotal` only
in comparison to the basket line items which are being decorated by both `price`
and `priceTotal`. The template relies on the models to be decorated by `price`
decorator and that is not the case.